Extending DP framework to support terms and stemm at the same time. This allows us to write optimizations based on similar context and clustering.
Added final methods to equals and hashCode in elements.
This class is thin wrapper arround Distance objects, but the equals method must be a comparisson of reference.
This allows us to groups lists with repeated elements.
Should not happen very often, but ate least the algorithm is robbust to this scenario.