Added a new method to identify knee in optimal curves.
This library already had Kneedle to find usefull bend points in
performance curves.
One drawback with this technique was long tails in the curve.
Since it uses the line y=x to find the minimal or maximum point in the
curve.
If the tail is long enought the maximum and minimum value will be
deslocated from the real value.
While trying to find a way to automatically cut the tail of the curve, I
develop a method based on line's slope.
The tail have slopes close to 0, since it is almost horizontal, while
the relevant part will have positive (or negative) slopes that are
bigger than zero in absolute value.

The algorithm identifies all the slopes between the points and uses
IsoData to select the ideal value that separates high slopes from small
ones.

Initial tests show that is works well on the Iris Dataset.