Added Rmethod to compute the Elbow and Knee points in a error curve.
The R in Rmethods stands for Regression.
It fits a power/ln function in the data (elbow/knee respectively). After that it computes the maximum value of the curvature Kf(x).

Most of the equations are already simplified, however the root of the final equations is estimated using the newton method.