Added methods to compute several Univariate regressions.
This code was on utils library, however the ArrayUtils library was becoming too big to handle.
As such, the regression part of the library is being added to this package.
As an improvement, the methods now return an Class that contains data regarding the regression (instead of a triplet of doubles).

The classes have implemented a solve method that computes values with the model. And a correct implementation of: toString, equals and hasdCode.