Fixed the issue with L-method and S-method.
As stated in the original paper, the iterative refinement stops when
the number of points is too small.
The code os S-method was improved, other performance optimizations are
being prepared.
The name of the S-method may be change in the future.
There are S shaped curves, that deserve the SMethod name.