Major refresh to the TM package.
1. Several classes got a build method.
It builds an instance with default paramerts and returns it.
Furthermore uses a WeakReference do the instance can be deleted after
usage.
It will be used to share important tools (stemmer, tokenizer,
clustering, etc...) it multiple processes/threads.
It also preventes that a single process to have several multiple
instances of a stemmer or tokenizer.

2. SLINK and CLINK can return clusters based on the elbow method.

3. DPW and DPWC have build method that use two caches to build new
profiles.

4. NGram has a new build named parse.
It uses a plain text tokenizer to build a n-gram.